Move full bolt DB implementation in boltdb.go
