+++ pool/acceptedwork.go
@@ -8,11 +8,7 @@ import (
 	"bytes"
 	"encoding/binary"
 	"encoding/hex"
-	"encoding/json"
-	"fmt"
 	"time"
-
-	bolt "go.etcd.io/bbolt"
 )
 
 // AcceptedWork represents an accepted work submission to the network.
@@ -58,176 +54,4 @@ func NewAcceptedWork(blockHash string, prevHash string, height uint32,
 		Miner:     miner,
 		CreatedOn: time.Now().UnixNano(),
 	}
-}
-
-// fetchAcceptedWork fetches the accepted work referenced by the provided id.
-// Returns an error if the work is not found.
-func (db *BoltDB) fetchAcceptedWork(id string) (*AcceptedWork, error) {
-	const funcName = "fetchAcceptedWork"
-	var work AcceptedWork
-	err := db.DB.View(func(tx *bolt.Tx) error {
-		bkt, err := fetchBucket(tx, workBkt)
-		if err != nil {
-			return err
-		}
-		v := bkt.Get([]byte(id))
-		if v == nil {
-			desc := fmt.Sprintf("%s: no value for key %s", funcName, id)
-			return dbError(ErrValueNotFound, desc)
-		}
-		err = json.Unmarshal(v, &work)
-		if err != nil {
-			desc := fmt.Sprintf("%s: unable to unmarshal accepted work: %v",
-				funcName, err)
-			return dbError(ErrParse, desc)
-		}
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-	return &work, err
-}
-
-// persistAcceptedWork saves the accepted work to the database.
-func (db *BoltDB) persistAcceptedWork(work *AcceptedWork) error {
-	const funcName = "persistAcceptedWork"
-	return db.DB.Update(func(tx *bolt.Tx) error {
-		bkt, err := fetchBucket(tx, workBkt)
-		if err != nil {
-			return err
-		}
-
-		// Do not persist already existing accepted work.
-		id := []byte(work.UUID)
-		v := bkt.Get(id)
-		if v != nil {
-			desc := fmt.Sprintf("%s: work %s already exists", funcName,
-				work.UUID)
-			return dbError(ErrValueFound, desc)
-		}
-		workBytes, err := json.Marshal(work)
-		if err != nil {
-			desc := fmt.Sprintf("%s: unable to marshal accepted "+
-				"work bytes: %v", funcName, err)
-			return dbError(ErrParse, desc)
-		}
-
-		err = bkt.Put(id, workBytes)
-		if err != nil {
-			desc := fmt.Sprintf("%s: unable to persist accepted work: %v",
-				funcName, err)
-			return dbError(ErrPersistEntry, desc)
-		}
-		return nil
-	})
-}
-
-// updateAcceptedWork persists modifications to an existing work. Returns an
-// error if the work is not found.
-func (db *BoltDB) updateAcceptedWork(work *AcceptedWork) error {
-	const funcName = "updateAcceptedWork"
-	return db.DB.Update(func(tx *bolt.Tx) error {
-		bkt, err := fetchBucket(tx, workBkt)
-		if err != nil {
-			return err
-		}
-
-		// Assert the work provided exists before updating.
-		id := []byte(work.UUID)
-		v := bkt.Get(id)
-		if v == nil {
-			desc := fmt.Sprintf("%s: work %s not found", funcName, work.UUID)
-			return dbError(ErrValueNotFound, desc)
-		}
-		workBytes, err := json.Marshal(work)
-		if err != nil {
-			desc := fmt.Sprintf("%s: unable to marshal accepted "+
-				"work bytes: %v", funcName, err)
-			return dbError(ErrPersistEntry, desc)
-		}
-		err = bkt.Put(id, workBytes)
-		if err != nil {
-			desc := fmt.Sprintf("%s: unable to persist accepted work: %v",
-				funcName, err)
-			return dbError(ErrPersistEntry, desc)
-		}
-		return nil
-	})
-}
-
-// deleteAcceptedWork removes the associated accepted work from the database.
-func (db *BoltDB) deleteAcceptedWork(id string) error {
-	return deleteEntry(db, workBkt, id)
-}
-
-// listMinedWork returns work data associated with all blocks mined by the pool
-// regardless of whether they are confirmed or not.
-//
-// List is ordered, most recent comes first.
-func (db *BoltDB) listMinedWork() ([]*AcceptedWork, error) {
-	const funcName = "listMinedWork"
-	minedWork := make([]*AcceptedWork, 0)
-	err := db.DB.View(func(tx *bolt.Tx) error {
-		bkt, err := fetchBucket(tx, workBkt)
-		if err != nil {
-			return err
-		}
-
-		cursor := bkt.Cursor()
-		for k, v := cursor.Last(); k != nil; k, v = cursor.Prev() {
-			var work AcceptedWork
-			err := json.Unmarshal(v, &work)
-			if err != nil {
-				desc := fmt.Sprintf("%s: unable to unmarshal accepted "+
-					"work: %v", funcName, err)
-				return poolError(ErrParse, desc)
-			}
-			minedWork = append(minedWork, &work)
-		}
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-	return minedWork, nil
-}
-
-// fetchUnconfirmedWork returns all work which is not confirmed as mined with
-// height less than the provided height.
-func (db *BoltDB) fetchUnconfirmedWork(height uint32) ([]*AcceptedWork, error) {
-	toReturn := make([]*AcceptedWork, 0)
-	err := db.DB.View(func(tx *bolt.Tx) error {
-		bkt, err := fetchBucket(tx, workBkt)
-		if err != nil {
-			return err
-		}
-
-		heightBE := heightToBigEndianBytes(height)
-		cursor := bkt.Cursor()
-		for k, v := cursor.First(); k != nil; k, v = cursor.Next() {
-			heightB, err := hex.DecodeString(string(k[:8]))
-			if err != nil {
-				return err
-			}
-
-			if bytes.Compare(heightBE, heightB) > 0 {
-				var work AcceptedWork
-				err := json.Unmarshal(v, &work)
-				if err != nil {
-					return err
-				}
-
-				if !work.Confirmed {
-					toReturn = append(toReturn, &work)
-				}
-			}
-		}
-
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-	return toReturn, nil
 }
