+++ pool/share.go
@@ -7,12 +7,8 @@ package pool
 import (
 	"bytes"
 	"encoding/hex"
-	"encoding/json"
-	"fmt"
 	"math/big"
 	"time"
-
-	bolt "go.etcd.io/bbolt"
 )
 
 // ShareWeights reprsents the associated weights for each known DCR miner.
@@ -54,180 +50,4 @@ func NewShare(account string, weight *big.Rat) *Share {
 		Weight:    weight,
 		CreatedOn: now,
 	}
-}
-
-// fetchShare fetches the share referenced by the provided id. Returns an error
-// if the share is not found.
-func (db *BoltDB) fetchShare(id string) (*Share, error) {
-	const funcName = "fetchShare"
-	var share Share
-	err := db.DB.View(func(tx *bolt.Tx) error {
-		bkt, err := fetchBucket(tx, shareBkt)
-		if err != nil {
-			return err
-		}
-
-		v := bkt.Get([]byte(id))
-		if v == nil {
-			desc := fmt.Sprintf("%s: no share found for id %s", funcName, id)
-			return dbError(ErrValueNotFound, desc)
-		}
-		err = json.Unmarshal(v, &share)
-		if err != nil {
-			desc := fmt.Sprintf("%s: unable to unmarshal share bytes: %v",
-				funcName, err)
-			return dbError(ErrParse, desc)
-		}
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-	return &share, err
-}
-
-// PersistShare saves a share to the database. Returns an error if a share
-// already exists with the same ID.
-func (db *BoltDB) PersistShare(s *Share) error {
-	const funcName = "PersistShare"
-	return db.DB.Update(func(tx *bolt.Tx) error {
-		bkt, err := fetchBucket(tx, shareBkt)
-		if err != nil {
-			return err
-		}
-
-		// Do not persist already existing share.
-		if bkt.Get([]byte(s.UUID)) != nil {
-			desc := fmt.Sprintf("%s: share %s already exists", funcName,
-				s.UUID)
-			return dbError(ErrValueFound, desc)
-		}
-
-		sBytes, err := json.Marshal(s)
-		if err != nil {
-			desc := fmt.Sprintf("%s: unable to marshal share bytes: %v",
-				funcName, err)
-			return dbError(ErrParse, desc)
-		}
-		err = bkt.Put([]byte(s.UUID), sBytes)
-		if err != nil {
-			desc := fmt.Sprintf("%s: unable to persist share entry: %v",
-				funcName, err)
-			return dbError(ErrPersistEntry, desc)
-		}
-		return nil
-	})
-}
-
-// ppsEligibleShares fetches all shares created before or at the provided time.
-func (db *BoltDB) ppsEligibleShares(max int64) ([]*Share, error) {
-	funcName := "ppsEligibleShares"
-	eligibleShares := make([]*Share, 0)
-	err := db.DB.View(func(tx *bolt.Tx) error {
-		bkt, err := fetchBucket(tx, shareBkt)
-		if err != nil {
-			return err
-		}
-		c := bkt.Cursor()
-		createdOnB := make([]byte, 8)
-		maxB := nanoToBigEndianBytes(max)
-		for k, v := c.First(); k != nil; k, v = c.Next() {
-			_, err := hex.Decode(createdOnB, k[:16])
-			if err != nil {
-				desc := fmt.Sprintf("%s: unable to decode share "+
-					"created-on bytes: %v", funcName, err)
-				return dbError(ErrDecode, desc)
-			}
-
-			if bytes.Compare(createdOnB, maxB) <= 0 {
-				var share Share
-				err := json.Unmarshal(v, &share)
-				if err != nil {
-					desc := fmt.Sprintf("%s: unable to unmarshal share: %v",
-						funcName, err)
-					return dbError(ErrParse, desc)
-				}
-				eligibleShares = append(eligibleShares, &share)
-			}
-		}
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-	return eligibleShares, err
-}
-
-// pplnsEligibleShares fetches all shares created after the provided time.
-func (db *BoltDB) pplnsEligibleShares(min int64) ([]*Share, error) {
-	funcName := "pplnsEligibleShares"
-	eligibleShares := make([]*Share, 0)
-	err := db.DB.View(func(tx *bolt.Tx) error {
-		bkt, err := fetchBucket(tx, shareBkt)
-		if err != nil {
-			return err
-		}
-		c := bkt.Cursor()
-		createdOnB := make([]byte, 8)
-		minB := nanoToBigEndianBytes(min)
-		for k, v := c.Last(); k != nil; k, v = c.Prev() {
-			_, err := hex.Decode(createdOnB, k[:16])
-			if err != nil {
-				desc := fmt.Sprintf("%s: unable to decode share "+
-					"created-on bytes: %v", funcName, err)
-				return dbError(ErrDecode, desc)
-			}
-
-			if bytes.Compare(createdOnB, minB) > 0 {
-				var share Share
-				err := json.Unmarshal(v, &share)
-				if err != nil {
-					desc := fmt.Sprintf("%s: unable to unmarshal "+
-						"share: %v", funcName, err)
-					return dbError(ErrParse, desc)
-				}
-				eligibleShares = append(eligibleShares, &share)
-			}
-		}
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-	return eligibleShares, err
-}
-
-// pruneShares removes shares with a createdOn time earlier than the provided
-// time.
-func (db *BoltDB) pruneShares(minNano int64) error {
-	funcName := "pruneShares"
-	minB := nanoToBigEndianBytes(minNano)
-
-	return db.DB.Update(func(tx *bolt.Tx) error {
-		bkt, err := fetchBucket(tx, shareBkt)
-		if err != nil {
-			return err
-		}
-		toDelete := [][]byte{}
-		cursor := bkt.Cursor()
-		createdOnB := make([]byte, 8)
-		for k, _ := cursor.First(); k != nil; k, _ = cursor.Next() {
-			_, err := hex.Decode(createdOnB, k[:16])
-			if err != nil {
-				desc := fmt.Sprintf("%s: unable to decode share created-on "+
-					"bytes: %v", funcName, err)
-				return dbError(ErrDecode, desc)
-			}
-			if bytes.Compare(minB, createdOnB) > 0 {
-				toDelete = append(toDelete, k)
-			}
-		}
-		for _, entry := range toDelete {
-			err := bkt.Delete(entry)
-			if err != nil {
-				return err
-			}
-		}
-		return nil
-	})
 }
